feat: add golden dataset evaluation pipeline for chunk retrieval
Implement comprehensive evaluation framework for assessing retrieval quality:

- LLM-assisted query generation with manual review workflow
- Structured rubric-based evaluation using LLM-as-judge pattern
- Metrics tracking: relevance, coverage, quality, and performance
- Aggregated reporting with detailed failure analysis

Components:
- QueryGenerator: Generate diverse evaluation queries
- RetrievalExecutor: Execute retrieval with timing
- ChunkQualityJudge: Evaluate using structured rubric
- EvaluationPipeline: Orchestrate full workflow

Features:
- Interactive CLI and programmatic API
- JSON-based query and result persistence
- Comprehensive metrics aggregation
- Human-readable reporting

Also includes:
- Fix temperature type from int to float in ModelConfig
- Add RetrieverSchema stub in handler.py

+++ src/elite_craft/model_provider.py
@@ -30,7 +30,7 @@ class ModelConfig:
         num_ctx: int = None,
         model_provider_url: str = None,
         reasoning: bool = False,
-        temperature: int = 0,
+        temperature: float = 0,
         use_ollama_local: bool = False,
         api_key: str = None,
         use_groq: bool = None,
